+++ scripts/_common/common.js
@@ -4,7 +4,25 @@ const parseValue = (v) => parseInt(String(v).replace(/[,\+%]+/g, '')); // will n
 const formatNumber = (v) => String(parseFloat(v).toFixed(2)).replace(/(\d)(?=(\d{3})+(?!\d))/g, "$1,"); // same format as the rest of the values in ui
 const formatNumberInt = (v) => String(Math.round(v)).replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1,'); // same format as the rest of the values in ui
 const toggleElement = (el, toggle) => { el.style = toggle ? 'display:block;' : 'display:none;'; };
-const mergeData = (model, data) => Object.keys(data || {}).forEach((key) => model[key] = data[key]);
+const ucfirst = (str) => String(str).charAt(0).toUpperCase() + String(str).slice(1);
+const lcfirst = (str) => String(str).charAt(0).toLowerCase() + String(str).slice(1);
+
+const pe = (v, c, s) => String(v).padEnd(c, s || ' ');
+const ps = (v, c, s) => String(v).padStart(c, s || ' ');
+
+const mergeData = (model, data, onlyExisting) => Object.keys(data || {}).forEach((key) => {
+    if (typeof model[key] === 'object' && model[key] !== null) {
+        if (typeof data[key] === 'object') {
+            mergeData(model[key], data[key], onlyExisting);
+        }
+    } else {
+        if (onlyExisting && !(key in model)) {
+            return;
+        }
+        model[key] = data[key];
+    }
+});
+
 const resetFilters = (filters, exclude) => {
     Array.from(filters).forEach((el) => {
         if (el !== exclude) {
@@ -19,25 +37,38 @@ const resetFilters = (filters, exclude) => {
     });
 };
 
-// dose not work because chrome blocks injecting js unless document changes the content policy
-// const injectScript = (source) => {
-//     const j = document.createElement('script'),
-//         f = document.getElementsByTagName('script')[0];
-//     j.textContent = source;
-//     f.parentNode.insertBefore(j, f);
-//     f.parentNode.removeChild(j);
-// };
+const getJsonPageData = () => {
+    let jsonPageData = {};
+    const jsPattern = /jsonPageData[\s]?=[\s]?(\{.*\});$/;
+    Array.from(document.querySelectorAll('script')).every((script) => {
+        if (jsPattern.test(script.innerHTML)) {
+            const [, jsonStr] = script.innerHTML.match(jsPattern);
+            jsonPageData = JSON.parse(jsonStr);
+            return false;
+        }
+        return true;
+    });
+    console.log('jsonPageData', jsonPageData);
+    return jsonPageData;
+}
 
-// const getJsonPageData = () => {
-//     function getJSvar(myvar) {
-//         document.body.setAttribute('data-' + myvar, JSON.stringify(myvar));
-//     }
-//     injectScript('(' + getJSvar.toString() + ')("jsonPageData")');
-//     return document.body.getAttribute('data-jsonPageData');
-// };
+const getTurnStatus = () => {
+    let turnStatus = {};
+    const jsPattern = /turnStatus[\s]?=[\s]?(\{.*\});/;
+    Array.from(document.querySelectorAll('script')).every((script) => {
+        if (jsPattern.test(script.innerHTML)) {
+            const [, jsonStr] = script.innerHTML.match(jsPattern);
+            turnStatus = JSON.parse(jsonStr);
+            return false;
+        }
+        return true;
+    });
+    console.log('turnStatus', turnStatus);
+    return turnStatus;
+}
 
-
-const currentTurn = () => document.querySelector('#turnNumber').innerText;
+//const currentTurn = () => document.querySelector('#turnNumber').innerText;
+const currentTurn = () => getTurnStatus().turnNumber;
 
 const schemeType = /class=\"(neutral|hostile|friendly|allied)\"/
 const getSchemeType = (el) => {