+++ scripts/scanpage/scan-processor.js
@@ -10,6 +10,10 @@ class ScanProcessorFactory {
                 return new SurfaceScanProcessor(turn);
                 break;
 
+            case ScanProcessor.TYPE_FLEET_SCAN:
+                return new FleetScanProcessor(turn);
+                break;
+
             default:
                 console.log('Unrecognised scan type', type);
                 return new ScanProcessor(turn);
@@ -56,6 +60,7 @@ class ScanProcessor {
     }
 
     parse(scanResult) {
+        console.log('scanResult', scanResult);
         this.planet = new dgPlanet(scanResult.coordinates.join('.') + ' ' + scanResult.name);
         mergeData(this.playerInfo, scanResult.playerInfo, true);
     }
@@ -144,8 +149,9 @@ class SurfaceScanProcessor extends ResourceScanProcessor {
     requiredSoldiersMax = 0;
     structures = [];
     summary = [];
+    extra = [];
     importantStructures = [
-        'Space_Tether',
+        41, // Space_Tether,
         'Hyperspace_Beacon',
         'Jump_Gate',
         39, // Comms_Satellite
@@ -174,13 +180,15 @@ class SurfaceScanProcessor extends ResourceScanProcessor {
             }
         });
         this.totals.worker += this.totals.occupiedWorker; // 'Worker' is in fact available workers
-        console.log('scanResult', scanResult, 'totals', this.totals, 'summary', this.summary);
         this.requiredSoldiers = this.getRequiredSoldiers(this.totals.worker, this.totals.soldier);
-        // @TODO add to housing
-        //this.requiredSoldiersMax = this.getRequiredSoldiers(this.houseingCapacity, this.totals.soldier);
         // @TODO add space ocupied by structures to total space (ground and orbit)
         //this.planetRating = new dgPlanetRating(this.stats); // with updated ground and orbit
 
+    }
+
+    setHouseingCapacity(houseingCapacity) {
+        this.houseingCapacity = houseingCapacity;
+        this.requiredSoldiersMax = this.getRequiredSoldiers(this.houseingCapacity, this.totals.soldier);
     }
 
     getRequiredSoldiers = (workers, soldiers) => Math.ceil((workers / 15) + (soldiers * 1.5) + 1);
@@ -197,6 +205,7 @@ class SurfaceScanProcessor extends ResourceScanProcessor {
             this.totals.worker,
             this.totals.soldier,
             this.requiredSoldiers,
+            this.requiredSoldiersMax,
             this.turn,
             '(' + this.totals.occupiedWorker + ' occupied workers) ' + this.summary.join(', '),
         ];
@@ -208,7 +217,7 @@ class SurfaceScanProcessor extends ResourceScanProcessor {
             ps(` Turn ${this.turn}`, 23, '='),
             this.planet.coords + ' ' + this.planet.name,
             `Owner: ${this.playerInfo.name} [${this.playerInfo.alliance.tag}]`,
-            ps('', 23, '-'),
+            pc(' ' + this.type + ' ', 23, '-'),
             pe('Metal:', 9) + ps(formatNumberInt(this.totals.metal), 14) + ' (' + this.stats.metal + '%)',
             pe('Mineral:', 9) + ps(formatNumberInt(this.totals.mineral), 14) + ' (' + this.stats.mineral + '%)',
             pe('Food:', 9) + ps(formatNumberInt(this.totals.food), 14) + ' (' + this.stats.food + '%)',
@@ -216,17 +225,61 @@ class SurfaceScanProcessor extends ResourceScanProcessor {
             ps('', 23, '-'),
             pe('Ground:', 14) + ps(this.stats.ground, 9),
             pe('Orbit:', 14) + ps(this.stats.orbit, 9),
-            pe('Rating:', 14) + ps(formatNumber(this.planetRating.rating.average), 9),
+            //pe('Rating:', 14) + ps(formatNumber(this.planetRating.rating.average), 9), // useless because th ground/oribt is not total
             ps('', 23, '-'),
             pe('Workers:', 14) + ps(formatNumberInt(this.totals.worker), 9),
             pe('Occupied:', 14) + ps(formatNumberInt(this.totals.occupiedWorker), 9),
             pe('Soldiers:', 14) + ps(formatNumberInt(this.totals.soldier), 9),
             pe('Required:', 14) + ps(formatNumberInt(this.requiredSoldiers), 9),
+            pe('Required max:', 14) + ps(formatNumberInt(this.requiredSoldiersMax), 9),
             ps('', 23, '-'),
             `Summary: ${this.summary.join(', ')}`,
             ps('', 23, '='),
         ];
         return "\n" + data.join("\n") + "\n";
     }
+}
 
+
+class FleetScanProcessor extends ScanProcessor {
+    fleetsByPlayer = {};
+    fleetsByAlliance = {};
+    fleetsByEta = {};
+
+    // indexes
+    ships = []; // all ships present in any fleet to build consistent summary
+    players = {};
+    alliances = {};
+
+    constructor(turn) {
+        super(turn);
+        this.type = ScanProcessor.TYPE_FLEET_SCAN;
+    }
+
+    parse(scanResult) {
+        super.parse(scanResult);
+        Array.from(scanResult.mobileList).forEach((item) => {
+            // console.log('item', item);
+            const alliance = item.player.alliance ? [
+                item.player.alliance.id,
+                '[' + item.player.alliance.tag + ']',
+                item.player.alliance.name,
+            ] : [];
+            console.log('ETA', item.moveTurns, 'name', item.name, 'total score', item.score, 'owner', item.player.id, item.player.name, 'alliance', alliance.join(' '));
+        });
+    }
+
+    addFleet(name, playerId, playerName, allianceTag, allianceName, eta, fleet) {
+        this.aliances[allianceTag] = allianceName;
+        this.players[playerId] = playerName;
+
+        this.fleetsByPlayer[playerId] || (this.fleetsByPlayer[playerId] = []);
+        this.fleetsByPlayer[playerId].push(fleet);
+
+        this.fleetsByAlliance[allianceTag] || (this.fleetsByAlliance[allianceTag] = []);
+        this.fleetsByAlliance[allianceTag].push(fleet);
+
+        this.fleetsByEta[eta] || (this.fleetsByEta[eta] = []);
+        this.fleetsByEta[eta].push(fleet);
+    }
 }